AHSniperToolV1/requirements.txt
Flask<1
flask-login
flask-script
flask-wtf
flask-sqlalchemy
pyexcel
pyexcel-xls
cx-Oracle
mysql-connector-python
authlib
requests